parsers/readme: A parsers for README files
There will be one single requirement named after the readme file.
all occurences of `=> REQ_ID` or `=> REQ_ID: title` will be added as
depend-reference to the readme requirment.

The Idea is that the README of a repository talks about the key points
of the project and links to the requirements that specify them in
detail.

The problem of all other toplevel requirements showing up as derived
when tracing the top level artefact agains a super-top-level
readme-artefact has to be tackled later on.